Issue #1404: Classes moved to acl package
