+++ api/src/main/java/com/epam/pipeline/controller/docker/DockerRegistryController.java
@@ -25,7 +25,7 @@ import com.epam.pipeline.entity.pipeline.DockerRegistryEventEnvelope;
 import com.epam.pipeline.entity.pipeline.Tool;
 import com.epam.pipeline.entity.security.JwtRawToken;
 import com.epam.pipeline.exception.docker.DockerAuthorizationException;
-import com.epam.pipeline.manager.docker.DockerRegistryApiService;
+import com.epam.pipeline.acl.docker.DockerRegistryApiService;
 import com.epam.pipeline.utils.AuthorizationUtils;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
